requirements.txt
torch==2.0.1+cu118
torchvision==0.15.2+cu118 --index-url https://download.pytorch.org/whl/cu118

transformers==4.34.0
open_clip_torch==2.23.0
ipykernel==6.25.1
ipython==8.12.2
ipywidgets==8.1.0
pandas==2.0.3
datasets==2.20.0
matplotlib==3.7.3
